Chapter6/5.LibrariesAndTransport/requirements.txt
-e ./adf
python-dotenv
uvicorn